buffer-pool/src/test/java/PoolArenaTest.java
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import pool.PoolArena;
import pool.PoolThreadCache;
import pool.PooledByteBuf;

import static org.junit.Assert.assertNotNull;
import static org.mockito.Mockito.atLeastOnce;
import static org.mockito.Mockito.verify;

public class PoolArenaTest {

    @Mock
    private PoolThreadCache mockCache;

    private PoolArena<Object> arena;

    @Before
    public void setUp() {
        MockitoAnnotations.openMocks(this);
        // 假设 PoolArena 有一个公共构造函数，用于实例化对象
        arena = new PoolArena<>(1024*8, 13, 1024*1024*16, 0);
    }

    @Test
    public void testAllocate() {
        int reqCapacity = 256;
        int maxCapacity = 1024;

        // 测试 allocate 方法的行为
        PooledByteBuf<Object> allocatedBuffer = arena.allocate(mockCache, reqCapacity, maxCapacity);

        // 校验返回的 PooledByteBuf 不为空
        assertNotNull("Allocated buffer should not be null", allocatedBuffer);
        // 这里添加更多校验来确认 allocatedBuffer 的状态是否符合预期

        // 验证 mockCache 是否在 allocate 方法中被正确使用
        PoolThreadCache verify = verify(mockCache, atLeastOnce());
        // 添加更多验证以确保方法的正确性
    }

}
